[EI-231] [api] refactor: mapper 함수 refactor 및 list 저장 로직 persistent adapter로 변경
