+++ api/src/numerical-guidance/infrastructure/adapter/twelve/mapper/live-indicator.mapper.ts
